Ensure family modal is always wired and restore banner
