+++ src/main.ts
@@ -43,7 +43,7 @@ import { recoveryText } from "@strings/recovery";
 import { mountMacToolbar, setAppToolbarTitle } from "@ui/AppToolbar";
 import { initAmbientBackground, type AmbientBackgroundController } from "@ui/AmbientBackground";
 import { getActiveTestScenario, getIpcAdapterName } from "@lib/ipc/provider";
-import { bannerFor } from "@ui/banner";
+import { updatePageBanner } from "@ui/updatePageBanner";
 
 // Resolve main app logo (SVG) as a URL the bundler can serve
 const appLogoUrl = new URL("./assets/logo.svg", import.meta.url).href;
@@ -419,53 +419,6 @@ function ensureCanonicalHash(route: RouteDefinition) {
     return;
   }
   history.replaceState(null, "", route.hash);
-}
-
-function updatePageBanner(route: RouteDefinition): void {
-  if (typeof document === "undefined") return;
-  const bannerEl = document.getElementById("page-banner") as HTMLDivElement | null;
-  if (!bannerEl) return;
-  const body = document.body;
-  const isFamilyInteractive = ENABLE_FAMILY_EXPANSION && route.id === "family";
-
-  if (isFamilyInteractive) {
-    bannerEl.hidden = false;
-    bannerEl.style.removeProperty("background-image");
-    bannerEl.style.removeProperty("--banner-pos-x");
-    bannerEl.style.removeProperty("--banner-pos-y");
-    bannerEl.setAttribute("aria-hidden", "false");
-    bannerEl.setAttribute("role", "complementary");
-    bannerEl.removeAttribute("aria-label");
-    bannerEl.dataset.bannerMode = "interactive";
-    body.dataset.bannerVisibility = "visible";
-    return;
-  }
-
-  if (bannerEl.dataset.bannerMode === "interactive") {
-    delete bannerEl.dataset.bannerMode;
-    bannerEl.innerHTML = "";
-    bannerEl.setAttribute("role", "img");
-  }
-  const key = route.id.toLowerCase();
-  const label = route.display?.label ?? key;
-  const url = bannerFor(key);
-  if (url) {
-    bannerEl.hidden = false;
-    bannerEl.style.backgroundImage = `url("${url}")`;
-    bannerEl.style.setProperty("--banner-pos-x", "50%");
-    bannerEl.style.setProperty("--banner-pos-y", "50%");
-    bannerEl.setAttribute("aria-hidden", "false");
-    bannerEl.setAttribute("aria-label", `${label} banner`);
-    body.dataset.bannerVisibility = "visible";
-  } else {
-    bannerEl.hidden = true;
-    bannerEl.style.removeProperty("background-image");
-    bannerEl.style.removeProperty("--banner-pos-x");
-    bannerEl.style.removeProperty("--banner-pos-y");
-    bannerEl.setAttribute("aria-hidden", "true");
-    bannerEl.removeAttribute("aria-label");
-    delete body.dataset.bannerVisibility;
-  }
 }
 
 async function renderApp({ route }: { route: RouteDefinition }) {
